editar solicitação

+++ api/app/Http/Controllers/CategoriesController.php
@@ -16,7 +16,8 @@ class CategoriesController extends Controller
 	public function data(Request $request, Category $category): JsonResponse
 	{
 		try {
-			$this->getPermissions($request->user());
+			$user = $this->getPermissions($request->user());
+			if ($user->category != 1) throw new \Exception('Não autorizado!');
 			return response()->json([
 				'status' => 'ok',
 				'message' => $category
@@ -32,7 +33,7 @@ class CategoriesController extends Controller
 	public function list(Request $request): JsonResponse
 	{
 		try {
-			$this->getPermissions($request->user());
+			$user = $this->getPermissions($request->user());
 			$categories = $this->repository->list($request->all());
 			$total = $this->repository->total($request->all());
 			$values = $this->getValues($total, $request->get('index'), $request->get('limit'));
@@ -55,7 +56,8 @@ class CategoriesController extends Controller
 	public function store(Request $request): JsonResponse
 	{
 		try {
-			$this->getPermissions($request->user());
+			$user = $this->getPermissions($request->user());
+			if ($user->category != 1) throw new \Exception('Não autorizado!');
 			$msg = $this->repository->add($request->all());
 			return response()->json([
 				'status' => 'ok',
@@ -72,7 +74,8 @@ class CategoriesController extends Controller
 	public function destroy(Request $request, int $id): JsonResponse
 	{
 		try {
-			$this->getPermissions($request->user());
+			$user = $this->getPermissions($request->user());
+			if ($user->category != 1) throw new \Exception('Não autorizado!');
 			return response()->json([
 				'status' => 'ok',
 				'message' => $this->repository->remove($id),
@@ -88,7 +91,8 @@ class CategoriesController extends Controller
 	public function change(Request $request, int $id): JsonResponse
 	{
 		try {
-			$this->getPermissions($request->user());
+			$user = $this->getPermissions($request->user());
+			if ($user->category != 1) throw new \Exception('Não autorizado!');
 			return response()->json([
 				'status' => 'ok',
 				'message' => $this->repository->change($id),
